Move SubmissionIndexCard to components
