+++ src/views/Home.js
@@ -1,24 +1,14 @@
 //@flow
 import React, { useEffect, useState } from "react"
 
-import Button from "@material-ui/core/Button"
-import Card from "@material-ui/core/Card"
-import CardActions from "@material-ui/core/CardActions"
-import CardContent from "@material-ui/core/CardContent"
-import CardHeader from "@material-ui/core/CardHeader"
 import CircularProgress from "@material-ui/core/CircularProgress"
 import Collapse from "@material-ui/core/Collapse"
 import Divider from "@material-ui/core/Divider"
 import Grid from "@material-ui/core/Grid"
-import List from "@material-ui/core/List"
-import ListItem from "@material-ui/core/ListItem"
-import ListItemIcon from "@material-ui/core/ListItemIcon"
-import ListItemText from "@material-ui/core/ListItemText"
 import { makeStyles } from "@material-ui/core/styles"
-import FolderIcon from "@material-ui/icons/Folder"
-import FolderOpenIcon from "@material-ui/icons/FolderOpen"
 import { useDispatch, useSelector } from "react-redux"
 
+import SubmissionIndexCard from "components/Home/SubmissionIndexCard"
 import WizardStatusMessageHandler from "components/NewDraftWizard/WizardForms/WizardStatusMessageHandler"
 import { setPublishedFolders } from "features/publishedFoldersSlice"
 import { setUnpublishedFolders } from "features/unpublishedFoldersSlice"
@@ -26,38 +16,8 @@ import { fetchUserById } from "features/userSlice"
 import folderAPIService from "services/folderAPI"
 
 const useStyles = makeStyles(theme => ({
-  cardGrid: {
-    paddingTop: theme.spacing(8),
-    paddingBottom: theme.spacing(8),
-  },
-  card: {
-    height: "100%",
-    display: "flex",
-    flexDirection: "column",
-    border: "none",
-    padding: theme.spacing(0),
-  },
-  cardTitle: {
-    fontSize: "0.5em",
-    padding: 0,
-    marginTop: theme.spacing(1),
-  },
-  cardContent: {
-    flexGrow: 1,
-    padding: 0,
-  },
   tableCard: {
     margin: theme.spacing(1, 0),
-  },
-  submissionsListItems: {
-    border: "solid 1px #ccc",
-    borderRadius: 3,
-    margin: theme.spacing(1, 0),
-    boxShadow: "0 1px 3px rgba(0,0,0,0.12), 0 1px 2px rgba(0,0,0,0.24)",
-    alignItems: "flex-start",
-  },
-  submissionsListIcon: {
-    minWidth: 35,
   },
   loggedUser: {
     margin: theme.spacing(2, 0, 0),
@@ -66,54 +26,6 @@ const useStyles = makeStyles(theme => ({
     margin: theme.spacing(10, "auto"),
   },
 }))
-
-type SubmissionIndexCardProps = {
-  folderType: string,
-  folderTitles: Array<string>,
-  buttonTitle: string,
-  onClickHeader: () => void,
-  onClickContent: () => void,
-  onClickButton: () => void,
-}
-
-const SubmissionIndexCard = (props: SubmissionIndexCardProps) => {
-  const classes = useStyles()
-  const { folderType, folderTitles, buttonTitle, onClickHeader, onClickContent, onClickButton } = props
-
-  return (
-    <Card className={classes.card} variant="outlined">
-      <CardHeader
-        title={folderType === "published" ? "Your Published Submissions" : "Your Draft Submissions"}
-        titleTypographyProps={{ variant: "subtitle1", fontWeight: "fontWeightBold" }}
-        className={classes.cardTitle}
-        onClick={onClickHeader}
-      />
-      <CardContent className={classes.cardContent} onClick={onClickContent}>
-        <List>
-          {folderTitles.map((folderTitle, index) => {
-            return (
-              <ListItem button key={index} dense className={classes.submissionsListItems}>
-                <ListItemIcon className={classes.submissionsListIcon}>
-                  {folderType === "published" ? <FolderIcon color="primary" /> : <FolderOpenIcon color="primary" />}
-                </ListItemIcon>
-                <ListItemText primary={folderTitle} />
-              </ListItem>
-            )
-          })}
-        </List>
-      </CardContent>
-      <CardActions>
-        {folderTitles.length > 0 && (
-          <Grid container alignItems="flex-start" justify="flex-end" direction="row">
-            <Button variant="outlined" color="primary" onClick={onClickButton}>
-              {buttonTitle}
-            </Button>
-          </Grid>
-        )}
-      </CardActions>
-    </Card>
-  )
-}
 
 const Home = () => {
   const dispatch = useDispatch()
